+++ app/code/ThirdParty/BlogArticle/Model/Resolver/DeleteBlogComment.php
@@ -1,0 +1,38 @@
+<?php
+namespace ThirdParty\BlogArticle\Model\Resolver;
+
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\GraphQl\Config\Element\Field;
+use Magento\Framework\GraphQl\Exception\GraphQlInputException;
+use Magento\Framework\GraphQl\Query\ResolverInterface;
+use Magento\Framework\GraphQl\Schema\Type\ResolveInfo;
+use ThirdParty\BlogArticle\Api\CommentRepositoryInterface;
+use ThirdParty\BlogArticle\Model\GraphQl\Authorization;
+
+class DeleteBlogComment implements ResolverInterface
+{
+    private $authorization;
+    private $commentRepository;
+
+    public function __construct(
+        Authorization $authorization,
+        CommentRepositoryInterface $commentRepository
+    ) {
+        $this->authorization = $authorization;
+        $this->commentRepository = $commentRepository;
+    }
+
+    public function resolve(Field $field, $context, ResolveInfo $info, ?array $value = null, ?array $args = null)
+    {
+        $this->authorization->assertCanWrite($context);
+        $commentId = isset($args['comment_id']) ? (int) $args['comment_id'] : 0;
+        if ($commentId <= 0) {
+            throw new GraphQlInputException(__('comment_id is required.'));
+        }
+        try {
+            return (bool) $this->commentRepository->deleteById($commentId);
+        } catch (LocalizedException $e) {
+            throw new GraphQlInputException(__($e->getMessage()), $e);
+        }
+    }
+}